pcform/requirements.txt
chardet==5.2.0
lxml==5.1.0
pillow==10.2.0
python-docx==1.1.0
pywin32==306
typing_extensions==4.10.0
